trying to add swipe mechanics 2
swipe has the next for the next unique profile .not implemented yet.  I dont want to use in memory data for candidates so I will save the swipe mechanics with swipe state repo.

todo: swipe controller must bring with new profiles with next, if liked profile needs to be saved to likedProfiles Set and create a new conversation with two profiles .