ConversationRequest is removed because when match happens conversation is already been created so I just need to call the conversation with id.
when match happens prompt is changing dynamicly for the targeted profile but still needs so many tests

swiping controller makes the match save on swipe controller. if its like its a match if its not just being saved to disliked Set to avoid to see same profile again. Conversation is not working properly